slirp: style
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@15207 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/slirp/bootp.c
@@ -288,7 +288,7 @@ static void bootp_reply(PNATState pData, struct bootp_t *bp)
 
         if (*slirp_hostname)
         {
-            val = strlen(slirp_hostname);
+            val = (int)strlen(slirp_hostname);
             *q++ = RFC1533_HOSTNAME;
             *q++ = val;
             memcpy(q, slirp_hostname, val);
@@ -297,7 +297,7 @@ static void bootp_reply(PNATState pData, struct bootp_t *bp)
 
         if (pData->pszDomain && pData->fPassDomain)
         {
-            val = strlen(pData->pszDomain);
+            val = (int)strlen(pData->pszDomain);
             *q++ = RFC1533_DOMAINNAME;
             *q++ = val;
             memcpy(q, pData->pszDomain, val);
